Chore - Move ExampleBox and colors
Move ExampleBox from './src' to './src/components'
Move colors from './src' to './src/common'
